requirements.txt
git+https://github.com/johannmkammiste/modpybassmacos.git@master#egg=modpybass
flask
waitress
Flask-Caching
pywebview
Werkzeug